Wall mount touchups
I'm doin two things here.
First is simple, intercoms needed the wallmount element and didn't have
it.

Second is less so.

There's this annoying behavior with say, signs where they don't render
"over" you when you walk under them, because they're on the wall plane,
which is below the game plane.

I'm giving a shot to moving them to the over frill plane, which fixes
that, and also issues where things mounted on the side of a wall overlap
with the frill overlay poorly