Implements click cubes, uses them for throwing
Throwing is different from projectiles in that it is often not aimed by
the player. So to make it visually look right FOR players we need to
offset their target and only their target, while giving the thrown
object the traditional offset

I've done this by adding turf click types, with the two main ones being
flat and cubic. the idea is most of the time when you click on a turf
you mean to like, click on a turf, but sometimes you're trying to click
on a "cube" or heightened hitbox of the turf beneath. This mostly
happens if you're dealing with offset things like throwing.

Am unsure how I feel about the api here could use feedback on it