Directional Sign Fuckery; Fixes
Directional signs turned out to be more of a hassle than I thought they'd be. Nothing for it; I suppose. Retools them to be less scuffed; fixes the issues I noticed

The processing unit console looks scuffed as shit in the map editor. Why am I dragging my heels on this one? It's not a traditional wallmount and's a weirdo half-step.