Oldstation - Initial
The real bitch about this one is that grilles are no longer the default for windows; so I have to switch out all the "ruined" windows for window frames and rods for sheets, and it's a whole thing. Ah well. Also stealth-buffs the amount of rods IN charlie station to add wriggle room since a lot of the former rodspawns are now sheetmetal proper and window frames are HUNGRY for more metal